server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
websockets==12.0
openai-whisper==20231117
numpy>=1.20.0
torch>=2.0.0
ffmpeg-python
python-jose==3.3.0 